[Step4] refactor : 코드 정리, 린트 적용
